requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
discord.py==1.4.0
idna==2.10
idna-ssl==1.1.0
multidict==4.7.6
typing-extensions==3.7.4.2
wincertstore==0.2
yarl==1.5.1
